Add selection for removal and replacement
