+++ storybook/src/stories/UseArray/Burger/assets/Cheese.tsx
