+++ storybook/src/stories/UseArray/Burger/assets/Ketchup.tsx
