+++ storybook/src/stories/UseArray/Burger/assets/Lettuce.tsx
