+++ storybook/src/stories/UseArray/Burger/assets/Mayo.tsx
