+++ storybook/src/stories/UseArray/Burger/assets/Patty.tsx
