+++ storybook/src/stories/UseArray/Burger/assets/Tomato.tsx
