+++ storybook/src/stories/UseArray/Burger/index.ts
@@ -1,0 +1,3 @@
+import Burger from './Burger'
+
+export default Burger
