+++ storybook/src/stories/UseArray/Burger/types.ts
@@ -1,0 +1,5 @@
+import { INGREDIENTS, INGREDIENT_GROUPS } from './constants'
+
+export type Ingredient = keyof typeof INGREDIENTS
+
+export type IngredientGroup = keyof typeof INGREDIENT_GROUPS
