+++ storybook/src/stories/UseArray/UseArray.stories.mdx
@@ -90,7 +90,15 @@ export declare type UseArrayReturn<T> = [
       newElement: React.SetStateAction<T>,
     ) => void
     /**
-     * The same functionality as JS Array.filter applied to state
+     * The same functionality as JS Array.map with the results applied to the array state
+     *
+     * @readonly
+     * @type {(newElement: React.SetStateAction<T>) => void}
+     * @param {React.SetStateAction<T>} newElement
+     */
+    readonly updateAll: (newElement: React.SetStateAction<T>) => void
+    /**
+     * The same functionality as JS Array.filter with the results applied to the array state
      *
      * @readonly
      * @type {(predicate: (element: T) => boolean) => void}
@@ -98,12 +106,18 @@ export declare type UseArrayReturn<T> = [
      */
     readonly filter: (predicate: (element: T) => boolean) => void
     /**
-     * The same functionality as JS Array.sort, but applied to state
+     * The same functionality as JS Array.sort with the results applied to the array state
      *
      * @type {((compareFn?: ((a: T, b: T) => number) | undefined) => void)}
      * @param {((a: T, b: T) => number) | undefined} compareFn
      */
     sort: (compareFn?: ((a: T, b: T) => number) | undefined) => void
+    /** /**
+     * The same functionality as JS Array.reverse with the results applied to the array state
+     *
+     * @type {() => void)}
+     */
+    reverse: () => void
     /**
      * Empties the array
      *
