+++ storybook/src/stories/UseArray/stories.tsx
@@ -1,10 +1,11 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import Burger from './Burger'
-import BURGER_CODE from './Burger.tsx?raw'
+import BURGER_CODE from './Burger/Burger.tsx?raw'
 
 export const burger: StoryObj<Meta<typeof Burger>> = {
   parameters: {
     layout: 'fullscreen',
+    backgrounds: { default: 'light', values: [{ name: 'light', value: 'white' }] },
     docs: {
       source: { code: BURGER_CODE, language: 'tsx' },
     },
